Modified the Multiple Instances policy to Parallel
Changed the Multiple Instances policy from IgnoreNew to Parallel

Parallel: Starts a new instance while an existing instance is running.
Queue: Starts a new instance of task after all other instances of the task are complete.
IgnoreNew: Default. Does not start a new instance if an existing instance of the task is running.
StopExisting: Stops an existing instance of the task before it starts a new instance.